feat: cria componnet do perfil
